+++ roomac_autonomous_manipulation/scripts/real_picking_object_manager.py
@@ -3,7 +3,7 @@
 
 import rospy
 from sensor_msgs.msg import JointState
-from geometry_msgs.msg import PointStamped
+from geometry_msgs.msg import PointStamped, Point
 
 from picking_object_manager import (
     PickingObjectManager,
@@ -20,6 +20,29 @@ class RealPickingObjectManager(PickingObjectManager):
             "/joint_states", JointState, queue_size=5
         )
 
+        object_type = "bottle"
+        self.object_point = Point()
+
+        # object is positioned away from ar tag + componsetion for
+        # difference between real manipulator and model
+
+        if object_type == "bottle":
+            self.object_point.point.x = -0.06 - 0.055
+            self.object_point.point.y = -0.16 - 0.008 + 0.04
+            self.object_point.point.z = 0.195 + 0.02
+        elif object_type == "cardbox_object":
+            self.object_point.point.x = -0.04 - 0.025
+            self.object_point.point.y = -0.09 - 0.008 + 0.02
+            self.object_point.point.z = 0.1
+
+        # Almost maximum, if needed can go to 1.2
+        self.opened_gripper_value = 1.0
+        self.closed_gripper_value = 0.1
+        if object_type == "bottle":
+            self.closed_gripper_value = 0.4
+        elif object_type == "cardbox_object":
+            self.closed_gripper_value = 0.1
+
     def move_gripper(self, position):
         gripper_msg = JointState()
         gripper_msg.name = ["right_gripper"]
@@ -30,24 +53,18 @@ class RealPickingObjectManager(PickingObjectManager):
 
     def close_gripper(self):
         rospy.loginfo("Sending close gripper command")
-        self.move_gripper(0.1)
+        self.move_gripper(self.closed_gripper_value)
 
     def open_gripper(self):
         rospy.loginfo("Sending open gripper command")
-        # Almost maximum, if needed can go to 1.2
-        self.move_gripper(1.0)
+        self.move_gripper(self.opened_gripper_value)
 
     def get_object_point(self):
-        object_point = PointStamped()
-        object_point.header.stamp = rospy.Time(0)
-        object_point.header.frame_id = "ar_marker_2"
-        # object is positioned away from ar tag + componsetion for
-        # difference between real manipulator and model
-        object_point.point.x = -0.04 - 0.025
-        object_point.point.y = -0.09 - 0.008 + 0.02
-        object_point.point.z = 0.1
-
-        return object_point
+        object_point_stamped = PointStamped()
+        object_point_stamped.header.stamp = rospy.Time(0)
+        object_point_stamped.header.frame_id = "detected_object"
+        object_point_stamped.point = self.object_point
+        return object_point_stamped
 
 
 if __name__ == "__main__":