+++ roomac_move_base/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package roomac_move_base
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.3.4 (2022-03-28)
+------------------
+
 1.3.3 (2022-03-26)
 ------------------
 * Merge pull request `#30 <https://github.com/macstepien/roomac_ros/issues/30>`_ from macstepien/teb-tuning